+++ handshake-nodejs-reporters/packages/handshake-dashboard/src/types/ui-constants.ts
@@ -1,0 +1,6 @@
+export const gridViewMode = 'grid';
+export const treeViewMode = 'tree';
+
+export const testEntitiesTab = 'test-entities';
+export const overviewTab = 'overview';
+export const timelineTab = 'timelineTab';