+++ handshake-nodejs-reporters/packages/wdio-handshake-reporter/src/index.ts
@@ -1,7 +1,7 @@
-import GraspItService from './service';
+import HandshakeService from './service';
 import HandshakeReporter from './reporter';
 
 export default HandshakeReporter;
 export * from './types';
-export const launcher = GraspItService;
+export const launcher = HandshakeService;
 export * from './helpers';